Integración de estilos y componentes de la vista
